src/agentic_platform/agent/pydanticai_agent/requirements.txt
pydantic-ai-slim[bedrock,openai]==0.2.14
openai==1.95.1
fastapi==0.115.12
uvicorn==0.34.0
boto3==1.37.27
pyjwt==2.10.1
cryptography==44.0.2
requests
litellm>=1.66.2
pydantic>=2.10.6